pyproject.toml
[project]
name = "AutoCoinMarginBot"
version = "0.1.0"
description = "Bybit anchored VWAP auto margin bot (PRD-driven scaffold)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "websockets>=12.0",
    "pydantic>=2.7",
    "pyyaml>=6.0",
    "loguru>=0.7",
]

[tool.setuptools]
packages = ["alerts", "clients", "config", "core", "strategy"]
